+++ NEWS.md
@@ -1,3 +1,8 @@
+# mlr3misc 0.1.7
+
+* `map_dtr()`, `imap_dtr` and `pmap_dtr` now pass `.idcol` down to argument
+  `idcol` of `data.table::rbindlist()`.
+
 # mlr3misc 0.1.6
 
 * `cite_bib()` or Rd macro `\cite{}` can now return the citation information of